requirements/requirements.txt
Flask==2.2.2
matplotlib==3.6.3
mediapipe==0.9.1.0
mysql_connector
numpy==1.24.2
opencv_python==4.7.0.68
Pillow==9.4.0
tensorflow-cpu==2.11.0
Werkzeug==2.2.2
yapf==0.40.1
pylint==2.17.5
